+++ components/amorphic/typescript_tests/apps/common/js/ticketItem.ts
@@ -4,7 +4,7 @@ import { Ticket } from './ticket';
 console.log('Compiling TicketItem');
 
 @supertypeClass
-export class TicketItem extends Persistable(Remoteable(Supertype())) {
+export class TicketItem extends Persistable(Remoteable(Supertype)) {
 
     // Secure properties can only be set on the server
     @property({getType: ()=>{return Person}})