+++ components/supertype/src/index.ts
@@ -18,8 +18,6 @@
  OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-
-import { Supertype } from '@haventech/supertype';
 import { ObjectTemplate } from './ObjectTemplate';
 
 ObjectTemplate.init();